+++ src/components/manage/Form/InlineForm.jsx
@@ -64,7 +64,7 @@ const InlineForm = ({
       )}
 
       <div id={`blockform-fieldset-${defaultFieldset.id}`}>
-        <Segment className="attached">
+        <Segment className="form">
           {map(defaultFieldset.fields, (field, index) => (
             <Field
               {...schema.properties[field]}
@@ -86,10 +86,8 @@ const InlineForm = ({
 
       {other.map((fieldset) => (
         <div key={fieldset.id} id={`blockform-fieldset-${fieldset.id}`}>
-          {title && (
-            <Segment className="secondary attached">{fieldset.title}</Segment>
-          )}
-          <Segment className="attached">
+          {title && <Segment className="secondary">{fieldset.title}</Segment>}
+          <Segment className="form">
             {map(fieldset.fields, (field) => (
               <Field
                 {...schema.properties[field]}